requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiohttp-swagger==1.0.16
aiosignal==1.2.0
arrow==1.2.2
async-lru==1.0.3
async-timeout==4.0.2
attrs==21.4.0
backports.ssl==0.0.9
bcrypt==3.2.2
binaryornot==0.4.4
certifi==2022.6.15
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.1.0
cisco-gnmi==1.0.15
click==8.1.3
cookiecutter==2.1.1
cryptography==37.0.2
dill==0.3.5.1
distro==1.7.0
frozenlist==1.3.0
genie==22.6
genie.libs.clean==22.6
genie.libs.conf==22.6
genie.libs.filetransferutils==22.6
genie.libs.health==22.6
genie.libs.ops==22.6
genie.libs.parser==22.6
genie.libs.robot==22.6
genie.libs.sdk==22.6
genie.telemetry==22.6
genie.trafficgen==22.6
gitdb==4.0.9
GitPython==3.1.18
grpcio==1.47.0
idna==3.3
IxNetwork==9.0.1915.16
ixnetwork-restpy==1.1.6
Jinja2==3.1.2
jinja2-time==0.2.0
jsonpickle==2.2.0
junit-xml==1.9
lxml==4.9.0
MarkupSafe==2.1.1
multidict==6.0.2
ncclient==0.6.13
netaddr==0.8.0
packaging==21.3
paramiko==2.11.0
pathspec==0.9.0
prettytable==3.3.0
protobuf==3.20.1
psutil==5.9.1
pyats==22.6
pyats.aereport==22.6
pyats.aetest==22.6
pyats.async==22.6
pyats.connections==22.6
pyats.contrib==22.6
pyats.datastructures==22.6
pyats.easypy==22.6
pyats.kleenex==22.6
pyats.log==22.6
pyats.reporter==22.6
pyats.results==22.6
pyats.robot==22.6
pyats.tcl==22.6
pyats.topology==22.6
pyats.utils==22.6
pycparser==2.21
pyftpdlib==1.5.6
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-engineio==3.14.2
python-slugify==6.1.2
python-socketio==4.6.1
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
robotframework==5.0.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
smmap==5.0.0
text-unidecode==1.3
tftpy==0.8.0
tqdm==4.64.0
unicon==22.6
unicon.plugins==22.6
urllib3==1.26.9
wcwidth==0.2.5
websocket-client==1.3.3
xlrd==1.2.0
XlsxWriter==3.0.3
xlwt==1.3.0
xmltodict==0.13.0
yamllint==1.26.3
yang.connector==22.6
yarl==1.7.2
